+++ src/hayhooks/server/utils/deploy_utils.py
@@ -1,14 +1,11 @@
 import asyncio
 import inspect
-import json
 import shutil
 import tempfile
 import threading
 import time
 import traceback
-from collections.abc import AsyncGenerator as AsyncGeneratorABC
 from collections.abc import Callable
-from collections.abc import Generator as GeneratorABC
 from functools import wraps
 from pathlib import Path
 from typing import Any
@@ -16,12 +13,10 @@ from typing import Any
 import docstring_parser
 from fastapi import FastAPI, Form, HTTPException
 from fastapi.concurrency import run_in_threadpool
-from fastapi.responses import Response, StreamingResponse
+from fastapi.responses import Response
 from fastapi.routing import APIRoute
-from haystack.dataclasses import StreamingChunk
 from pydantic import BaseModel
 
-from hayhooks.events import PipelineEvent
 from hayhooks.server.exceptions import PipelineAlreadyExistsError, PipelineFilesError
 from hayhooks.server.logger import log, log_elapsed
 from hayhooks.server.pipelines import registry
@@ -30,7 +25,6 @@ from hayhooks.server.pipelines.models import (
     create_response_model_from_callable,
     get_response_class_from_callable,
 )
-from hayhooks.server.pipelines.sse import SSEStream
 from hayhooks.server.utils.base_pipeline_wrapper import BasePipelineWrapper
 from hayhooks.server.utils.models import PreparedPipeline
 from hayhooks.server.utils.module_loader import (
@@ -38,6 +32,7 @@ from hayhooks.server.utils.module_loader import (
     load_pipeline_module,
     unload_pipeline_modules,
 )
+from hayhooks.server.utils.streaming_response_utils import _streaming_response_from_result
 from hayhooks.server.utils.yaml_pipeline_wrapper import YAMLPipelineWrapper
 from hayhooks.settings import DeployConcurrencyPolicy, settings
 
@@ -219,111 +214,6 @@ def handle_pipeline_exceptions() -> Callable:
         return wrapper
 
     return decorator
-
-
-def _format_run_stream_chunk(stream_item: Any) -> str | bytes | None:
-    if isinstance(stream_item, StreamingChunk):
-        if stream_item.content:
-            return stream_item.content
-        reasoning = getattr(stream_item, "reasoning", None)
-        if reasoning is not None:
-            return getattr(reasoning, "reasoning_text", "") or ""
-        return ""
-    if isinstance(stream_item, PipelineEvent):
-        log.warning("PipelineEvent emitted during /run streaming; skipping. Use OpenAI chat endpoints for UI events.")
-        return None
-    if isinstance(stream_item, (str, bytes)):
-        return stream_item
-    if stream_item is None:
-        return ""
-    try:
-        return json.dumps(stream_item)
-    except TypeError:
-        return str(stream_item)
-
-
-def _format_sse_chunk(formatted: str | bytes) -> str:
-    text = formatted.decode("utf-8", errors="replace") if isinstance(formatted, bytes) else str(formatted)
-
-    if text == "":
-        return "data:\n\n"
-
-    lines = text.splitlines()
-    if not lines:
-        return "data:\n\n"
-
-    data_lines = "".join(f"data: {line}\n" for line in lines)
-    return f"{data_lines}\n"
-
-
-def _streaming_response_from_async_gen(async_gen: Any, media_type: str = "text/plain") -> Response:
-    is_sse = media_type == "text/event-stream"
-
-    async def async_stream():
-        try:
-            async for item in async_gen:
-                formatted = _format_run_stream_chunk(item)
-                if formatted is None:
-                    continue
-                if is_sse:
-                    formatted = _format_sse_chunk(formatted)
-                yield formatted
-        finally:
-            aclose = getattr(async_gen, "aclose", None)
-            if callable(aclose):
-                await aclose()
-
-    return StreamingResponse(async_stream(), media_type=media_type)
-
-
-def _streaming_response_from_gen(gen: Any, media_type: str = "text/plain") -> Response:
-    is_sse = media_type == "text/event-stream"
-
-    def sync_stream():
-        try:
-            for item in gen:
-                formatted = _format_run_stream_chunk(item)
-                if formatted is None:
-                    continue
-                if is_sse:
-                    formatted = _format_sse_chunk(formatted)
-                yield formatted
-        finally:
-            close = getattr(gen, "close", None)
-            if callable(close):
-                close()
-
-    return StreamingResponse(sync_stream(), media_type=media_type)
-
-
-def _streaming_response_from_result(result: Any) -> Response | None:
-    # If the result is a SSEStream, return a StreamingResponse with the appropriate media type
-    if isinstance(result, SSEStream):
-        # Get the stream from the SSEStream
-        stream = result.stream
-
-        # If the stream is an async generator, return a StreamingResponse with the appropriate media type
-        if isinstance(stream, AsyncGeneratorABC):
-            return _streaming_response_from_async_gen(stream, media_type="text/event-stream")
-
-        # If the stream is a generator, return a StreamingResponse with the appropriate media type
-        if isinstance(stream, GeneratorABC):
-            return _streaming_response_from_gen(stream, media_type="text/event-stream")
-
-        # If the stream is not a generator or async generator, raise a TypeError
-        msg = f"SSEStream.stream must be a generator or async generator (got type {type(stream)!r})"
-        raise TypeError(msg)
-
-    # If the result is a Response, return the result
-    if isinstance(result, Response):
-        return result
-
-    # Following generic cases are for non-SSE streaming responses (plain text)
-    if isinstance(result, AsyncGeneratorABC):
-        return _streaming_response_from_async_gen(result)
-    if isinstance(result, GeneratorABC):
-        return _streaming_response_from_gen(result)
-    return None
 
 
 async def _execute_pipeline_run(
